// src/types/TProductSizeGuide.tsx
export type TProductSizeGuide = {
    id: number
    productId: number
    size?: string
    length?: number
    chest?: number
    shoulder?: number
    sleeveLength?: number
    sleeveOpening?: number
    frontCrotch?: number
    waist?: number
    armhole?: number
    bottom?: number
    sortOrder?: number
}